Better handle websocket interrupts, and preventing high cpu usage
