+++ Cargo.toml
@@ -17,6 +17,7 @@ bcrypt = "0.10"
 rand = "0.8.4"
 chrono = "0.4.19"
 crossbeam-channel = "0.5.1"
+crossbeam = "0.8.1"
 hyper = { version = "0.14.9", features = ["full"] }
 tokio = { version = "1.7.1", features = ["full"] }
 hyper-tungstenite = "0.3.3"
